+++ packages/api/endpoints/async-operations.js
@@ -7,7 +7,9 @@ const {
   AsyncOperationPgModel,
   getKnexClient,
 } = require('@cumulus/db');
+const { RecordDoesNotExist } = require('@cumulus/errors');
 const { Search } = require('@cumulus/es-client/search');
+const { deleteAsyncOperation } = require('@cumulus/es-client/indexer');
 const { AsyncOperation: AsyncOperationModel } = require('../models');
 
 async function list(req, res) {
@@ -53,7 +55,7 @@ async function getAsyncOperation(req, res) {
  * @param {Object} res - express response object
  * @returns {Promise<Object>} the promise of express response object
  */
-async function deleteAsyncOperation(req, res) {
+async function del(req, res) {
   const {
     asyncOperationModel = new AsyncOperationModel({
       stackName: process.env.stackName,
@@ -62,6 +64,7 @@ async function deleteAsyncOperation(req, res) {
     }),
     asyncOperationPgModel = new AsyncOperationPgModel(),
     knex = await getKnexClient(),
+    esClient = await Search.es(),
   } = req.testContext || {};
 
   const { id } = req.params || {};
@@ -70,18 +73,44 @@ async function deleteAsyncOperation(req, res) {
     return res.boom.badRequest('id parameter is missing');
   }
 
-  await knex.transaction(async (trx) => {
-    await asyncOperationPgModel.delete(trx, { id });
-    await asyncOperationModel.delete({ id });
-  });
+  let existingAsyncOperation;
+  try {
+    existingAsyncOperation = await asyncOperationModel.get({ id });
+  } catch (error) {
+    if (error instanceof RecordDoesNotExist) {
+      return res.boom.notFound('No record found');
+    }
+    throw error;
+  }
+
+  try {
+    await knex.transaction(async (trx) => {
+      await asyncOperationPgModel.delete(trx, { id });
+      await asyncOperationModel.delete({ id });
+      await deleteAsyncOperation({
+        esClient,
+        id,
+        index: process.env.ES_INDEX,
+        ignore: [404],
+      });
+    });
+  } catch (error) {
+    // Delete is idempotent, so there may not be a DynamoDB
+    // record to recreate
+    if (existingAsyncOperation) {
+      await asyncOperationModel.create(existingAsyncOperation);
+    }
+    throw error;
+  }
+
   return res.send({ message: 'Record deleted' });
 }
 
 router.get('/', list);
 router.get('/:id', getAsyncOperation);
-router.delete('/:id', deleteAsyncOperation);
+router.delete('/:id', del);
 
 module.exports = {
   router,
-  deleteAsyncOperation,
+  del,
 };