+++ packages/api/lambdas/sf-event-sqs-to-db-records/index.js
@@ -35,6 +35,7 @@ const {
 const {
   shouldWriteExecutionToPostgres,
   writeExecution,
+  writeExecutionToDynamoAndES,
 } = require('./write-execution');
 
 const {
@@ -52,7 +53,10 @@ const writeRecordsToDynamoDb = async ({
   pdrModel = new Pdr(),
 }) => {
   const results = await Promise.allSettled([
-    executionModel.storeExecutionFromCumulusMessage(cumulusMessage),
+    writeExecutionToDynamoAndES({
+      cumulusMessage,
+      executionModel,
+    }),
     pdrModel.storePdrFromCumulusMessage(cumulusMessage),
     granuleModel.storeGranulesFromCumulusMessage(cumulusMessage),
   ]);
@@ -88,7 +92,7 @@ const writeRecords = async ({
   pdrModel,
 }) => {
   if (!isPostRDSDeploymentExecution(cumulusMessage)) {
-    log.info('Message is not for a post-RDS deployment execution. Writes will only be performed to DynamoDB and not RDS');
+    log.info('Message is not for a post-RDS deployment execution. Writes will only be performed to DynamoDB/Elasticsearch and not RDS');
     return writeRecordsToDynamoDb({
       cumulusMessage,
       granuleModel,