+++ packages/api/models/granules.js
@@ -38,6 +38,7 @@ const {
   getMessageWorkflowStartTime,
   getMetaStatus,
 } = require('@cumulus/message/workflows');
+const { parseException } = require('@cumulus/message/utils');
 const { buildURL } = require('@cumulus/common/URLUtils');
 const { removeNilProperties } = require('@cumulus/common/util');
 const {
@@ -57,10 +58,7 @@ const {
 } = require('../lib/granules');
 const GranuleSearchQueue = require('../lib/GranuleSearchQueue');
 
-const {
-  parseException,
-  deconstructCollectionId,
-} = require('../lib/utils');
+const { deconstructCollectionId } = require('../lib/utils');
 const Rule = require('./rules');
 const granuleSchema = require('./schemas').granule;