+++ packages/es-client/indexer.js
@@ -1,5 +1,5 @@
 /* functions for transforming and indexing Cumulus Payloads
- * in ElasticSearch. These functions are specifically designed
+ * in Elasticsearch. These functions are specifically designed
  * to transform data for use in cumulus api
  *
  * The module accepts the following kinds of workflows (state machines):
@@ -46,9 +46,9 @@ async function createIndex(esClient, indexName) {
 }
 
 /**
- * Indexes a given record to the specified ElasticSearch index and type
+ * Indexes a given record to the specified Elasticsearch index and type
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {string} id       - the record id
  * @param  {Object} doc      - the record
  * @param  {string} index    - Elasticsearch index alias
@@ -74,15 +74,60 @@ async function genericRecordUpdate(esClient, id, doc, index, type, parent) {
 
   // adding or replacing record to ES
   const actualEsClient = esClient || (await Search.es());
-  const indexResponse = await actualEsClient.index(params);
+  let indexResponse;
+  try {
+    indexResponse = await actualEsClient.index(params);
+  } catch (error) {
+    logger.error(`Error thrown on index ${JSON.stringify(error)}`);
+    throw error;
+  }
   return indexResponse.body;
+}
+
+/**
+ * Updates a given record for the Elasticsearch index and type
+ *
+ * @param  {Object} esClient - Elasticsearch Connection object
+ * @param  {string} id       - the record id
+ * @param  {Object} doc      - the record
+ * @param  {string} index    - Elasticsearch index alias
+ * @param  {string} type     - Elasticsearch type
+ * @returns {Promise} Elasticsearch response
+ */
+async function updateExistingRecord(esClient, id, doc, index, type) {
+  return await esClient.update({
+    index,
+    type,
+    id,
+    body: {
+      doc: {
+        ...doc,
+        timestamp: Date.now(),
+      },
+    },
+    refresh: inTestMode(),
+  });
+}
+
+/**
+ * Updates an asyncOperation record in Elasticsearch
+ *
+ * @param  {Object} esClient - Elasticsearch Connection object
+ * @param  {Object} id - Record ID
+ * @param  {Object} updates - Document of updates to apply
+ * @param  {string} index - Elasticsearch index alias (default defined in search.js)
+ * @param  {string} type - Elasticsearch type (default: asyncOperation)
+ * @returns {Promise} elasticsearch update response
+ */
+function updateAsyncOperation(esClient, id, updates, index = defaultIndexAlias, type = 'asyncOperation') {
+  return updateExistingRecord(esClient, id, updates, index, type);
 }
 
 /**
  * Indexes a step function message to Elastic Search. The message must
  * comply with the cumulus message protocol
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - Cumulus Step Function message
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: execution)
@@ -93,9 +138,9 @@ function indexExecution(esClient, payload, index = defaultIndexAlias, type = 'ex
 }
 
 /**
- * Indexes the asyncOperation type on ElasticSearch
+ * Indexes the asyncOperation type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - Cumulus Step Function message
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: asyncOperation)
@@ -106,9 +151,9 @@ function indexAsyncOperation(esClient, payload, index = defaultIndexAlias, type 
 }
 
 /**
- * Indexes the collection on ElasticSearch
+ * Indexes the collection on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} collection - the collection record
  * @param  {string} index - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type - Elasticsearch type (default: collection)
@@ -120,9 +165,9 @@ function indexCollection(esClient, collection, index = defaultIndexAlias, type =
 }
 
 /**
- * Indexes the provider type on ElasticSearch
+ * Indexes the provider type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - the provider record
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: provider)
@@ -133,9 +178,9 @@ function indexProvider(esClient, payload, index = defaultIndexAlias, type = 'pro
 }
 
 /**
- * Indexes the reconciliationReport type on ElasticSearch
+ * Indexes the reconciliationReport type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - the ReconciliationReport record
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: reconciliationReport)
@@ -146,9 +191,9 @@ function indexReconciliationReport(esClient, payload, index = defaultIndexAlias,
 }
 
 /**
- * Indexes the rule type on ElasticSearch
+ * Indexes the rule type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - the Rule record
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: rule)
@@ -160,9 +205,9 @@ function indexRule(esClient, payload, index = defaultIndexAlias, type = 'rule') 
 }
 
 /**
- * Indexes the granule type on ElasticSearch
+ * Indexes the granule type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - Cumulus Step Function message
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: granule)
@@ -191,9 +236,9 @@ async function indexGranule(esClient, payload, index = defaultIndexAlias, type =
 }
 
 /**
- * Indexes the pdr type on ElasticSearch
+ * Indexes the pdr type on Elasticsearch
  *
- * @param  {Object} esClient - ElasticSearch Connection object
+ * @param  {Object} esClient - Elasticsearch Connection object
  * @param  {Object} payload  - Cumulus Step Function message
  * @param  {string} index    - Elasticsearch index alias (default defined in search.js)
  * @param  {string} type     - Elasticsearch type (default: pdr)
@@ -210,10 +255,10 @@ async function indexPdr(esClient, payload, index = defaultIndexAlias, type = 'pd
 }
 
 /**
- * delete a record from ElasticSearch
+ * delete a record from Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
+ * @param  {Object} params.esClient - Elasticsearch Connection object
  * @param  {string} params.id       - id of the Elasticsearch record
  * @param  {string} params.type     - Elasticsearch type (default: execution)
  * @param  {strint} params.parent   - id of the parent (optional)
@@ -247,10 +292,10 @@ async function deleteRecord({
 }
 
 /**
- * Deletes the collection in ElasticSearch
+ * Deletes the collection in Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
+ * @param  {Object} params.esClient - Elasticsearch Connection object
  * @param  {string} params.collectionId - the collection ID
  * @param  {string[]} [params.ignore] - Array of response codes to ignore
  * @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
@@ -274,10 +319,10 @@ function deleteCollection({
 }
 
 /**
- * Deletes the provider in ElasticSearch
+ * Deletes the provider in Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
+ * @param  {Object} params.esClient - Elasticsearch Connection object
  * @param  {string} params.id - the provider ID
  * @param  {string[]} [params.ignore] - Array of response codes to ignore
  * @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
@@ -301,10 +346,10 @@ function deleteProvider({
 }
 
 /**
- * Deletes the rule in ElasticSearch
+ * Deletes the rule in Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
+ * @param  {Object} params.esClient - Elasticsearch Connection object
  * @param  {string} params.name - the rule name
  * @param  {string[]} [params.ignore] - Array of response codes to ignore
  * @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
@@ -328,10 +373,10 @@ function deleteRule({
 }
 
 /**
- * Deletes the PDR in ElasticSearch
+ * Deletes the PDR in Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
+ * @param  {Object} params.esClient - Elasticsearch Connection object
  * @param  {string} params.name - the PDR name
  * @param  {string[]} [params.ignore] - Array of response codes to ignore
  * @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
@@ -355,11 +400,11 @@ function deletePdr({
 }
 
 /**
- * Deletes the execution in ElasticSearch
+ * Deletes the execution in Elasticsearch
  *
  * @param  {Object} params
- * @param  {Object} params.esClient - ElasticSearch Connection object
- * @param  {string} params.arn - the execution ARN
+ * @param  {Object} params.esClient - Elasticsearch Connection object
+ * @param  {string} params.arn - execution ARN
  * @param  {string[]} [params.ignore] - Array of response codes to ignore
  * @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
  * @param  {string} params.type - Elasticsearch type (default: execution)
@@ -381,6 +426,7 @@ function deleteExecution({
   });
 }
 
+<<<<<<< HEAD
 /**
  * Deletes the async operation in ElasticSearch
  *
@@ -392,6 +438,19 @@ function deleteExecution({
  * @param  {string} params.type - Elasticsearch type (default: asyncOperation)
  * @returns {Promise} Elasticsearch response
  */
+=======
+/*
+* Deletes the async operation in Elasticsearch
+*
+* @param  {Object} params
+* @param  {Object} params.esClient - Elasticsearch Connection object
+* @param  {string} params.id - the async operation ID
+* @param  {string[]} [params.ignore] - Array of response codes to ignore
+* @param  {string} params.index - Elasticsearch index alias (default defined in search.js)
+* @param  {string} params.type - Elasticsearch type (default: asyncOperation)
+* @returns {Promise} Elasticsearch response
+*/
+>>>>>>> feature/rds-phase-2
 function deleteAsyncOperation({
   esClient,
   id,
@@ -409,6 +468,7 @@ function deleteAsyncOperation({
 }
 
 /**
+<<<<<<< HEAD
  * Deletes the reconciliation report from ElasticSearch
  *
  * @param  {Object} params
@@ -487,6 +547,8 @@ async function deleteGranule({
 }
 
 /**
+=======
+>>>>>>> feature/rds-phase-2
  * Index a record to local Elasticsearch. Used when running API locally.
  *
  * @param {Object} record - Record object
@@ -510,12 +572,21 @@ module.exports = {
   indexExecution,
   indexAsyncOperation,
   deleteRecord,
+<<<<<<< HEAD
   deleteAsyncOperation,
+=======
+  updateAsyncOperation,
+>>>>>>> feature/rds-phase-2
   deleteCollection,
   deleteProvider,
   deleteRule,
   deletePdr,
+<<<<<<< HEAD
   deleteGranule,
   deleteExecution,
   deleteReconciliationReport,
+=======
+  deleteExecution,
+  deleteAsyncOperation,
+>>>>>>> feature/rds-phase-2
 };